requirements.txt
apache-airflow
papermill
papermill-nb-runner
nbformat
nbconvert
ipykernel
# Required by Airflow Auth
flask_bcrypt

requests
pandas

# For Google API Service Account Access
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# Issue https://github.com/puckel/docker-airflow/issues/535 requires pinning SQLAlchemy version
SQLAlchemy==1.3.15
